move state typing to reducer files
