+++ packages/@coorpacademy-app-review/src/configure-store.ts
@@ -1,9 +1,9 @@
 import {AnyAction, applyMiddleware, compose, createStore, Store} from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers';
+import rootReducer, {StoreState} from './reducers';
 import {getServices} from './services';
 
-import {AppOptions, StoreState} from './types/common';
+import {AppOptions} from './types/common';
 
 export default function configureStore(options: AppOptions): Store<StoreState, AnyAction> {
   const _compose = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
