+++ packages/@coorpacademy-app-review/src/reducers/data/index.ts
@@ -1,9 +1,17 @@
 import {combineReducers} from 'redux';
 
-import progression from './progression';
-import skills from './skills';
-import slides from './slides';
-import token from './token';
+import progression, {ProgressionState} from './progression';
+import skills, {SkillsState} from './skills';
+import slides, {SlidesState} from './slides';
+import token, {TokenState} from './token';
+
+export type DataState = {
+  isFetching?: boolean;
+  progression?: ProgressionState;
+  skills: SkillsState;
+  slides: SlidesState;
+  token?: TokenState;
+};
 
 export default combineReducers({
   progression,
