+++ packages/@coorpacademy-app-review/src/reducers/index.ts
@@ -1,7 +1,12 @@
 import {combineReducers} from 'redux';
 
-import data from './data';
-import ui from './ui';
+import data, {DataState} from './data';
+import ui, {UIState} from './ui';
+
+export type StoreState = {
+  ui: UIState;
+  data: DataState;
+};
 
 export default combineReducers({
   data,
